Fixes #29, Fix memory leaks for commands in main.c
Commands are held in an array and has options to control the identity
for the command and the function to call for it.
Add main.h that holds the macros and types for main.c.

Remove parameter name from forward declarations.

Don't cast size_t to int, when not needed.