Add file configuration.
Add --config option to read a file other than the default user specific
config file.

If a --config option is not given read a config file which is ignored if
it doesn't exist.
On Windows the config file is in %APPDATA%/.sphiarc
On Linux/BSD/etc the cofnig file resides at $HOME/.sphiarc

SPHIA_PATH is given the top priority if given, instead of only being used
if no other path option is set.